Move the searching of book cover to addBook function
if addBook is called with a book with a missing cover, it will
asynchronously try to fetch a missing cover image from an api.
it will also temporarily change the cover of the book to a "searching
for cover" image.

this commit also fixes the problem of the app retrying to fetch an image
when it was unsuccessful in the first attempt. this was happening because
there was no way to know a previous search had failed. now if a cover
image is not found, it sets the appropriate "missing cover image"
message and a search for that particular book cover is not made in the future.